docs: added responsive Buttons to view the components
Here the possibility was implemented to view the components in their responsive representation. For
this purpose the components are displayed one by one on a page which can be found under
"/raw-component/{:component-name}". This page is integrated on the detail page of the component via
iFrame. To change the view, buttons have been added which influence the width of the displayed
iFrame.