fix-googlemaps fix: switching languages
Fixed an issue where the googlemaps api would break if the language was changed on a page that includes the Google Maps Section without a page reload. The Google Maps Section now manually loads the googlemaps api instead of using the loader from google. That way we can easily delete the api on the page before reloading it with a different language.