fix(newsdetail component): changed some props and layout issues
To provide a better user experience some properties has been changed. The image now uses the actual
image props, the date is not a date anymore but a string which is directy displayed and there is a
new way to display the social icons with a slot which can be used

BREAKING CHANGE: The property date has been changed from Date to string. The property socialText
does not exist anymore and is transformed an object.