fix-googlemaps fix(googlemaps section): changed the way the startLocation is handled
The start location is now a mandatory prop. This is because the fallback to a default behaviour by
google does not work as desired. So the start location is now the fallback in case the user denies
the request for permission to use their location. Also added a boolean to enforce the start location
without prompting the user for their location. In case this behaviour is desired.