fix-googlemaps fix(googlemaps section): fixed an issue with closing info windows
When closing an info window using its 'x' button the selectedIndex would not be set to null. This
would cause a click on the same marker to not register and open the window again because it was
still the selected item. This issue is now fixed in case the window gets closed using the 'x'
button. The index wil still not be set to null if the window gets closed through some other means.