TNG-809-fix-image fix(image component): using an illegal value for "darken" (fka "opacity") now throws an error
The opacity prop is now named "darken" and a validation method was added. It should throw an error
if an illegal value is used. And should accept strings and numbers (if the component is being used
in a javascript environment). If an error is thrown, it will still result in a black image. Also
updated the documentation.