Fix googlemaps (#69)
* ci(semantic-release): enable next, alpha and beta branches for automatic releases

* Update release.yml

ci (semantic-release): Allow release on other branches as well.

* fix-googlemaps
docs(googlemaps section): added link to the allowed language strings to the description

Removed check for 2 character language strings. Removed the corresponding unit test. Added a link to
the documentation that contains a table of all allowed language strings.

* fix-googlemaps
docs(googlemaps section): added a mention that the surrounding element needs a height

* fix-googlemaps
fix(googlemaps section): the description box should no longer contain the words undefined

When a location property was missing, which could only happen if the component was rendered in a
javascript environment, the description box would render 'undefined' instead. This should no longer
happen. If a property is missing the corresponding HTML Element will not be rendered at all.

* fix-googlemaps
fix(googlemaps section): changed the way the startLocation is handled

The start location is now a mandatory prop. This is because the fallback to a default behaviour by
google does not work as desired. So the start location is now the fallback in case the user denies
the request for permission to use their location. Also added a boolean to enforce the start location
without prompting the user for their location. In case this behaviour is desired.

* fix-googlemaps
fix(googlemaps section): the sidebar should disappear now when no locations exists

* fix-googlemaps
docs(googlemaps section): added example for renderInfoWindow prop

* fix-googlemaps
docs(googlemaps section): added example that showcases the rest of the props

* fix-googlemaps
fix(googlemaps section): fixed an issue with closing info windows

When closing an info window using its 'x' button the selectedIndex would not be set to null. This
would cause a click on the same marker to not register and open the window again because it was
still the selected item. This issue is now fixed in case the window gets closed using the 'x'
button. The index wil still not be set to null if the window gets closed through some other means.

* fix-googlemaps
fix(googlemaps section): language is no longer required. Default language is now navigator.language

* fix-googlemaps
docs(documentation meta component): meta comp. now sets document.title to its own title when mounted

* add 'fsxa-ui' to page title

Co-authored-by: Tobias Golbs <37590108+tgolbs@users.noreply.github.com>
Co-authored-by: Emre Neumann <neumann@e-spirit.com>
Co-authored-by: Vlado Plaga <plaga@e-spirit.com>